Fix demo structure: move runTensor to src, add scalar/tensor mode selection
