+++ packages/demo/package.json
@@ -7,6 +7,8 @@
   },  
   "scripts": {
     "build": "tsc -b .",
-    "start": "pnpm build && node dist/index.js"
+    "start": "pnpm build && node dist/index.js",
+    "start:tensor": "pnpm build && node dist/index.js tensor",
+    "start:scalar": "pnpm build && node dist/index.js scalar"
   }
 }
